refactor(generator): extract EndpointFinder from SymbolDiscovery
